Update slice.lua to more pythonic slice style
First, it was a huge mistake: with index -1 we got mapped index #array -1 but not #array. In other words, when we have
a = {10, 20, 30}
The slice calcs index -1 as 20, not 30. 

Second, out-of-bound indexes weren't clamped. 

Third, negative indexes didn't work:
For a = {10, 20, 30, 40, 50}
If we want the last 3 elements and do
slice(a, -1, -3)
we'll got left = 4, right = 2, left > right and our loop doesn't run.